crewAI/requirements.txt
crewai
crewai-tools
mistralai
python-dotenv
elevenlabs
